+++ apps/code-examples/src/app/features/layout.module.ts
@@ -7,6 +7,12 @@ import { BoxDemoComponent as InlineHelpBoxDemoComponent } from '../code-examples
 import { BoxDemoModule as InlineHelpBoxDemoModule } from '../code-examples/layout/box/inline-help/box-demo.module';
 import { DescriptionListDemoComponent } from '../code-examples/layout/description-list/horizontal/description-list-demo.component';
 import { DescriptionListDemoModule } from '../code-examples/layout/description-list/horizontal/description-list-demo.module';
+import { FormatDemoComponent } from '../code-examples/layout/format/format-demo.component';
+import { FormatDemoModule } from '../code-examples/layout/format/format-demo.module';
+import { TextExpandRepeaterDemoComponent } from '../code-examples/layout/text-expand-repeater/text-expand-repeater-demo.component';
+import { TextExpandRepeaterDemoModule } from '../code-examples/layout/text-expand-repeater/text-expand-repeater-demo.module';
+import { TextExpandDemoComponent } from '../code-examples/layout/text-expand/inline/text-expand-demo.component';
+import { TextExpandDemoModule } from '../code-examples/layout/text-expand/inline/text-expand-demo.module';
 
 const routes: Routes = [
   {
@@ -21,6 +27,18 @@ const routes: Routes = [
     path: 'description-list',
     component: DescriptionListDemoComponent,
   },
+  {
+    path: 'format',
+    component: FormatDemoComponent,
+  },
+  {
+    path: 'text-expand-repeater',
+    component: TextExpandRepeaterDemoComponent,
+  },
+  {
+    path: 'text-expand',
+    component: TextExpandDemoComponent,
+  },
 ];
 
 @NgModule({
@@ -34,7 +52,10 @@ export class LayoutRoutingModule {}
     BoxDemoModule,
     InlineHelpBoxDemoModule,
     DescriptionListDemoModule,
+    FormatDemoModule,
     LayoutRoutingModule,
+    TextExpandRepeaterDemoModule,
+    TextExpandDemoModule,
   ],
 })
 export class LayoutModule {}