+++ libs/components/layout/src/lib/modules/action-button/action-button-adapter-service.spec.ts
@@ -12,13 +12,13 @@ import { SkyActionButtonAdapterService } from './action-button-adapter-service';
   `,
 })
 class SkyActionButtonAdapterTestComponent {
-  public parentWidth: string;
+  public parentWidth: string | undefined;
 
   @ViewChild('el', {
     read: ElementRef,
     static: true,
   })
-  public input: ElementRef;
+  public input!: ElementRef;
 }
 
 describe('Action button adapter service', () => {