+++ libs/components/layout/src/lib/modules/page-summary/page-summary-adapter.service.ts
@@ -5,7 +5,7 @@ import { ElementRef, Injectable } from '@angular/core';
  */
 @Injectable()
 export class SkyPageSummaryAdapterService {
-  public updateKeyInfoLocation(elRef: ElementRef, isXS: boolean) {
+  public updateKeyInfoLocation(elRef: ElementRef, isXS: boolean): void {
     const el = elRef.nativeElement;
     const keyInfoContainerEl = el.querySelector(
       '.sky-page-summary-key-info-container'