+++ libs/components/layout/src/lib/modules/text-expand/text-expand.component.ts
@@ -1,4 +1,11 @@
 import {
+  animate,
+  state,
+  style,
+  transition,
+  trigger,
+} from '@angular/animations';
+import {
   AfterContentInit,
   Component,
   ElementRef,
@@ -12,7 +19,7 @@ import { forkJoin as observableForkJoin } from 'rxjs';
 import { take } from 'rxjs/operators';
 
 import { SkyTextExpandAdapterService } from './text-expand-adapter.service';
-import { SkyTextExpandModalContext } from './text-expand-modal-context';
+import { SKY_TEXT_EXPAND_MODAL_CONTEXT } from './text-expand-modal-context-token';
 import { SkyTextExpandModalComponent } from './text-expand-modal.component';
 
 /**
@@ -21,6 +28,28 @@ import { SkyTextExpandModalComponent } from './text-expand-modal.component';
 let nextId = 0;
 
 @Component({
+  animations: [
+    trigger('expansionAnimation', [
+      transition(':enter', []),
+      state(
+        'true',
+        style({
+          maxHeight: '{{transitionHeight}}px',
+        }),
+        { params: { transitionHeight: 0 } }
+      ),
+      state(
+        'false',
+        style({
+          maxHeight: '{{transitionHeight}}px',
+        }),
+        { params: { transitionHeight: 0 } }
+      ),
+      transition('true => false', animate('250ms ease')),
+      transition('false => true', animate('250ms ease')),
+      transition('void => *', []),
+    ]),
+  ],
   selector: 'sky-text-expand',
   templateUrl: './text-expand.component.html',
   styleUrls: ['./text-expand.component.scss'],
@@ -32,23 +61,47 @@ export class SkyTextExpandComponent implements AfterContentInit {
    * @default "Expanded view"
    */
   @Input()
-  public expandModalTitle: string;
+  public expandModalTitle: string | undefined;
 
   /**
    * Specifies the maximum number of text characters to display inline when users select the link
    * to expand the full text. If the text exceeds this limit, then the component expands
    * the full text in a modal instead.
+   * @default 600
    */
   @Input()
-  public maxExpandedLength = 600;
+  public set maxExpandedLength(value: number | undefined) {
+    if (value) {
+      this.#_maxExpandedLength = value;
+    } else {
+      this.#_maxExpandedLength = 600;
+    }
+    this.#setup(this.text);
+  }
+
+  public get maxExpandedLength(): number {
+    return this.#_maxExpandedLength;
+  }
 
   /**
    * Specifies the maximum number of newline characters to display inline when users select
    * the link to expand the full text. If the text exceeds this limit, then
    * the component expands the full text in a modal view instead.
+   * @default 2
    */
   @Input()
-  public maxExpandedNewlines = 2;
+  public set maxExpandedNewlines(value: number | undefined) {
+    if (value) {
+      this.#_maxExpandedNewlines = value;
+    } else {
+      this.#_maxExpandedNewlines = 2;
+    }
+    this.#setup(this.text);
+  }
+
+  public get maxExpandedNewlines(): number {
+    return this.#_maxExpandedNewlines;
+  }
 
   /**
    * Specifies the number of text characters to display before truncating the text.
@@ -57,28 +110,30 @@ export class SkyTextExpandComponent implements AfterContentInit {
    * @default 200
    */
   @Input()
-  public set maxLength(value: number) {
-    this._maxLength = value;
-
-    /* istanbul ignore else */
-    if (this.textEl) {
-      this.setup(this.expandedText);
+  public set maxLength(value: number | undefined) {
+    if (value) {
+      this.#_maxLength = value;
+    } else {
+      this.#_maxLength = 200;
     }
+    this.#setup(this.text);
   }
 
   public get maxLength(): number {
-    return this._maxLength;
+    return this.#_maxLength;
   }
 
   /**
    * Specifies the text to truncate.
    */
   @Input()
-  public set text(value: string) {
-    /* istanbul ignore else */
-    if (this.textEl) {
-      this.setup(value);
-    }
+  public set text(value: string | undefined) {
+    this.#_text = value ?? '';
+    this.#setup(this.#_text);
+  }
+
+  public get text(): string {
+    return this.#_text;
   }
 
   /**
@@ -87,47 +142,70 @@ export class SkyTextExpandComponent implements AfterContentInit {
   @Input()
   public truncateNewlines = true;
 
-  public buttonText: string;
+  public buttonText = '';
 
   public contentSectionId = `sky-text-expand-content-${++nextId}`;
 
-  public expandable: boolean;
+  public expandable = false;
 
-  public isExpanded = false;
+  public isExpanded: boolean | undefined;
 
-  public isModal: boolean;
+  public isModal = false;
+
+  public transitionHeight = 1;
 
   @ViewChild('container', {
     read: ElementRef,
     static: true,
   })
-  private containerEl: ElementRef;
+  public containerEl: ElementRef | undefined;
 
   @ViewChild('text', {
     read: ElementRef,
     static: true,
   })
-  private textEl: ElementRef;
+  public set textEl(value: ElementRef | undefined) {
+    this.#_textEl = value;
+    this.#setup(this.text);
+  }
 
-  private collapsedText: string;
+  public get textEl(): ElementRef | undefined {
+    return this.#_textEl;
+  }
 
-  private expandedText: string;
+  #collapsedText = '';
 
-  private newlineCount: number;
+  #newlineCount = 0;
 
-  private seeMoreText: string;
+  #seeMoreText = '';
 
-  private seeLessText: string;
+  #seeLessText = '';
 
-  private textToShow: string;
+  #textToShow = '';
 
-  private _maxLength = 200;
+  #_maxExpandedLength = 600;
+
+  #_maxExpandedNewlines = 2;
+
+  #_maxLength = 200;
+
+  #_text = '';
+
+  #_textEl: ElementRef | undefined;
+
+  #resources: SkyLibResourcesService;
+  #modalSvc: SkyModalService;
+  #textExpandAdapter: SkyTextExpandAdapterService;
 
   constructor(
-    private resources: SkyLibResourcesService,
-    private modalService: SkyModalService,
-    private textExpandAdapter: SkyTextExpandAdapterService
-  ) {}
+    resources: SkyLibResourcesService,
+    modalSvc: SkyModalService,
+    textExpandAdapter: SkyTextExpandAdapterService
+  ) {
+    this.#resources = resources;
+    this.#modalSvc = modalSvc;
+    this.#textExpandAdapter = textExpandAdapter;
+  }
 
   public textExpand(): void {
     if (this.isModal) {
@@ -135,12 +213,12 @@ export class SkyTextExpandComponent implements AfterContentInit {
       /* istanbul ignore else */
       /* sanity check */
       if (!this.isExpanded) {
-        this.modalService.open(SkyTextExpandModalComponent, [
+        this.#modalSvc.open(SkyTextExpandModalComponent, [
           {
-            provide: SkyTextExpandModalContext,
+            provide: SKY_TEXT_EXPAND_MODAL_CONTEXT,
             useValue: {
               header: this.expandModalTitle,
-              text: this.expandedText,
+              text: this.text,
             },
           },
         ]);
@@ -148,42 +226,41 @@ export class SkyTextExpandComponent implements AfterContentInit {
     } else {
       // Normal View
       if (!this.isExpanded) {
-        this.setContainerMaxHeight();
-        setTimeout(() => {
-          this.isExpanded = true;
-          this.animateText(this.collapsedText, this.expandedText, true);
-        }, 10);
+        this.#animateText(true);
       } else {
-        this.setContainerMaxHeight();
-        setTimeout(() => {
-          this.isExpanded = false;
-          this.animateText(this.expandedText, this.collapsedText, false);
-        }, 10);
+        this.#animateText(false);
       }
     }
   }
 
   public animationEnd(): void {
-    // Ensure the correct text is displayed
-    this.textExpandAdapter.setText(this.textEl, this.textToShow);
-    // Set height back to auto so the browser can change the height as needed with window changes
-    this.textExpandAdapter.setContainerHeight(this.containerEl, undefined);
+    if (this.textEl && this.containerEl) {
+      // Ensure the correct text is displayed
+      this.#textExpandAdapter.setText(this.textEl, this.#textToShow);
+
+      setTimeout(() => {
+        if (this.containerEl) {
+          // Set height back to auto so the browser can change the height as needed with window changes
+          this.#textExpandAdapter.removeContainerMaxHeight(this.containerEl);
+        }
+      });
+    }
   }
 
   public ngAfterContentInit(): void {
     observableForkJoin([
-      this.resources.getString('skyux_text_expand_see_more'),
-      this.resources.getString('skyux_text_expand_see_less'),
+      this.#resources.getString('skyux_text_expand_see_more'),
+      this.#resources.getString('skyux_text_expand_see_less'),
     ])
       .pipe(take(1))
       .subscribe((resources) => {
-        this.seeMoreText = resources[0];
-        this.seeLessText = resources[1];
-        this.setup(this.expandedText);
+        this.#seeMoreText = resources[0];
+        this.#seeLessText = resources[1];
+        this.#setup(this.text);
 
         /* istanbul ignore else */
         if (!this.expandModalTitle) {
-          this.resources
+          this.#resources
             .getString('skyux_text_expand_modal_title')
             .pipe(take(1))
             .subscribe((resource) => {
@@ -193,44 +270,31 @@ export class SkyTextExpandComponent implements AfterContentInit {
       });
   }
 
-  private setContainerMaxHeight(): void {
-    // ensure everything is reset
-    this.animationEnd();
-    /* Before animation is kicked off, ensure that a maxHeight exists */
-    /* Once we have support for angular v4 animations with parameters we can use that instead */
-    const currentHeight = this.textExpandAdapter.getContainerHeight(
-      this.containerEl
-    );
-    this.textExpandAdapter.setContainerHeight(
-      this.containerEl,
-      `${currentHeight}px`
-    );
-  }
-
-  private setup(value: string): void {
+  #setup(value: string | undefined): void {
     if (value) {
-      this.newlineCount = this.getNewlineCount(value);
-      this.collapsedText = this.getTruncatedText(value, this.maxLength);
-      this.expandedText = value;
-      if (this.collapsedText !== value) {
-        this.buttonText = this.seeMoreText;
+      this.#newlineCount = this.#getNewlineCount(value);
+      this.#collapsedText = this.#getTruncatedText(value, this.maxLength);
+      if (this.#collapsedText !== value) {
+        this.buttonText = this.#seeMoreText;
         this.isExpanded = false;
         this.expandable = true;
         this.isModal =
-          this.newlineCount > this.maxExpandedNewlines ||
-          this.expandedText.length > this.maxExpandedLength;
+          this.#newlineCount > this.maxExpandedNewlines ||
+          this.text.length > this.maxExpandedLength;
       } else {
         this.expandable = false;
       }
-      this.textToShow = this.collapsedText;
+      this.#textToShow = this.#collapsedText;
     } else {
-      this.textToShow = '';
+      this.#textToShow = '';
       this.expandable = false;
     }
-    this.textExpandAdapter.setText(this.textEl, this.textToShow);
+    if (this.textEl) {
+      this.#textExpandAdapter.setText(this.textEl, this.#textToShow);
+    }
   }
 
-  private getNewlineCount(value: string): number {
+  #getNewlineCount(value: string): number {
     const matches = value.match(/\n/gi);
 
     if (matches) {
@@ -240,7 +304,7 @@ export class SkyTextExpandComponent implements AfterContentInit {
     return 0;
   }
 
-  private getTruncatedText(value: string, length: number): string {
+  #getTruncatedText(value: string, length: number): string {
     let i: number;
     if (this.truncateNewlines) {
       value = value.replace(/\n+/gi, ' ');
@@ -256,41 +320,29 @@ export class SkyTextExpandComponent implements AfterContentInit {
         }
       }
     }
-    return value.substr(0, length);
+    return value.substring(0, length);
   }
 
-  private animateText(
-    previousText: string,
-    newText: string,
-    expanding: boolean
-  ): void {
-    const adapter = this.textExpandAdapter;
-    const container = this.containerEl;
-    // Reset max height
-    adapter.setContainerHeight(container, undefined);
-    // Measure the current height so we can animate from it.
-    const currentHeight = adapter.getContainerHeight(container);
-    this.textToShow = newText;
-    adapter.setText(this.textEl, this.textToShow);
-    this.buttonText = expanding ? this.seeLessText : this.seeMoreText;
-    // Measure the new height so we can animate to it.
-    const newHeight = adapter.getContainerHeight(container);
-    /* istanbul ignore if */
-    if (newHeight < currentHeight) {
-      // The new text is smaller than the old text, so put the old text back before doing
-      // the collapse animation to avoid showing a big chunk of whitespace.
-      adapter.setText(this.textEl, previousText);
+  #animateText(expanding: boolean): void {
+    if (this.containerEl && this.textEl) {
+      const adapter = this.#textExpandAdapter;
+      const container = this.containerEl;
+      if (expanding) {
+        adapter.setText(this.textEl, this.text);
+        this.#textToShow = this.text;
+      } else {
+        adapter.setText(this.textEl, this.#collapsedText);
+        this.#textToShow = this.#collapsedText;
+      }
+      this.buttonText = expanding ? this.#seeLessText : this.#seeMoreText;
+      // Measure the new height so we can animate to it.
+      const newHeight = adapter.getContainerHeight(container);
+      this.transitionHeight = newHeight;
+      // Always show all text while animating so that the animation is smooth. The animation callback will set this back if needed.
+      if (!expanding) {
+        adapter.setText(this.textEl, this.text);
+      }
+      this.isExpanded = expanding;
     }
-
-    adapter.setContainerHeight(container, `${currentHeight}px`);
-    // This timeout is necessary due to the browser needing to pick up the non-auto height being set
-    // in order to do the transtion in height correctly. Without it the transition does not fire.
-    setTimeout(() => {
-      adapter.setContainerHeight(container, `${newHeight}px`);
-      /* This resets values if the transition does not get kicked off */
-      setTimeout(() => {
-        this.animationEnd();
-      }, 500);
-    }, 10);
   }
 }