+++ libs/components/modals/src/lib/modules/modal/modal.component.spec.ts
@@ -30,6 +30,7 @@ import { SkyModalBeforeCloseHandler } from './modal-before-close-handler';
 import { SkyModalComponentAdapterService } from './modal-component-adapter.service';
 import { SkyModalHostService } from './modal-host.service';
 import { SkyModalInstance } from './modal-instance';
+import { SkyModalScrollShadowDirective } from './modal-scroll-shadow.directive';
 import { SkyModalService } from './modal.service';
 
 describe('Modal component', () => {
@@ -104,10 +105,6 @@ describe('Modal component', () => {
 
   function getRouter(): Router {
     return TestBed.inject(Router);
-  }
-
-  function getMockMutationObserverService(): ModalMockMutationObserverService {
-    return TestBed.inject<any>(SkyMutationObserverService);
   }
 
   function openModal<T>(modalType: T, config?: Record<string, any>) {
@@ -901,8 +898,21 @@ describe('Modal component', () => {
     }
 
     beforeEach(() => {
+      const modalMockMutationObserverService =
+        new ModalMockMutationObserverService();
+
+      TestBed.overrideDirective(SkyModalScrollShadowDirective, {
+        add: {
+          providers: [
+            {
+              provide: MutationObserverService,
+              useValue: modalMockMutationObserverService,
+            },
+          ],
+        },
+      });
       mutationObserverCreateSpy = spyOn(
-        getMockMutationObserverService(),
+        modalMockMutationObserverService,
         'create'
       ).and.callThrough();